% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tour-utils.R
\name{clamp}
\alias{clamp}
\title{Rescale all columns to lie in unit interval}
\usage{
clamp(.data)
}
\arguments{
\item{.data}{A 'matrix' like object}
}
\description{
Rescale all columns to lie in unit interval
}
\details{
This function "clamps" all columns in \code{.data} to fit
onto the \code{ncol(.data)} dimensional unit cube.
}
