% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tour-utils.R
\name{clamp_robust}
\alias{clamp_robust}
\title{Rescale all columns by median and IQR}
\usage{
clamp_robust(.data)
}
\arguments{
\item{.data}{A 'matrix' like object}
}
\description{
Rescale all columns by median and IQR
}
