% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tour-utils.R
\name{clamp_sd}
\alias{clamp_sd}
\title{Rescale all columns to have same standard deviation}
\usage{
clamp_sd(.data, sd = 1e-04)
}
\arguments{
\item{.data}{A 'matrix' like object}

\item{sd}{the value of each columns standard devation}
}
\description{
Rescale all columns to have same standard deviation
}
