% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tour-utils.R
\name{compute_half_range}
\alias{compute_half_range}
\title{Compute range of axes for bases}
\usage{
compute_half_range(.data)
}
\arguments{
\item{.data}{A 'matrix' like object}
}
\description{
Compute range of axes for bases
}
\details{
This function computes the maximum squared
Euclidean distance of rows in a matrix like object. Mostly used
internally for setting up xy-axis ranges for a tour animation.
}
