% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tour-utils.R
\name{compute_proj_dist}
\alias{compute_proj_dist}
\title{Compute Frobenius norm of matrix-like objects x and y}
\usage{
compute_proj_dist(x, y)
}
\arguments{
\item{x, y}{'matrix' like objects that have \code{tcrossprod} methods}
}
\value{
A numeric vector of length 1 that is the Frobenius norm
}
\description{
Compute Frobenius norm of matrix-like objects x and y
}
\examples{
x <- matrix(rnorm(300), ncol = 3)
y <- matrix(rnorm(300), ncol = 3)
compute_proj_dist(x, y)
}
