% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liminal-data.R
\docType{data}
\name{fake_trees}
\alias{fake_trees}
\title{A high-dimensional tree data structure with 10 branching points.}
\format{
An object of class \code{data.frame} with 3000 rows and 101 columns.
}
\source{
\href{https://github.com/KrishnaswamyLab/PHATE/blob/master/Python/phate/tree.py}{PHATE}
}
\usage{
fake_trees
}
\description{
A high-dimensional tree data structure with 10 branching points.
}
\details{
Data are obtained from diffusion limited aggregation
tree simulation in the \code{phate} python and \code{phateR} packages, but
reconstructed as a wide data.frame rather than a list.

There are 3000 rows and 101 columns, the first 100 columns are labelled
dim1 - dim100, and are numeric, while the final column is a
factor representing the branch id.
}
\keyword{datasets}
