% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbours.R
\name{find_knn}
\alias{find_knn}
\title{Utilities for estimating  and searching kNN graph}
\usage{
find_knn(
  .data,
  num_neighbors = 10,
  metric = c("euclidean", "cosine", "manhattan"),
  .include_self = TRUE
)
}
\description{
Utilities for estimating  and searching kNN graph
}
