% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embed.R
\name{limn_embed_tsne}
\alias{limn_embed_tsne}
\title{Generate multiple t-SNE embeddings for a single data set}
\usage{
limn_embed_tsne(.data, cols, .params, ...)
}
\arguments{
\item{.data}{a data.frame to extract tSNE embeddings for}

\item{cols}{columns to embed}

\item{.params}{A named list with valid parameters to the embedding function,
this will expand out all combinations of inputs}

\item{...}{other paramters to pass to \code{\link[Rtsne:Rtsne]{Rtsne::Rtsne()}}}
}
\description{
Generate multiple t-SNE embeddings for a single data set
}
