% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter.R
\name{limn_xy}
\alias{limn_xy}
\title{Brushed Scatter Plot}
\usage{
limn_xy(
  .data,
  x,
  y,
  color = NULL,
  brush_name = "brush",
  brush_transform = NULL,
  brush_transform_mark = "square",
  ...
)
}
\arguments{
\item{.data}{input data.frame}

\item{x, y}{bare symbols or expressions using columns from \code{.data} to map
to the x and y aesthetics of a scatter plot}

\item{color}{an optional sybmol or expression using a column from \code{.data} to
map to the color aesthetic of a scatter plot.}

\item{brush_name}{the name of the brush element, default is 'brush'}

\item{brush_transform}{aggregate data inside the brush (currently supports
median / mean), default is NULL.}

\item{brush_transform_mark}{how to display result of brush transform
(default is "square")}

\item{...}{other parameters to passed to \code{\link[vegawidget:vegawidget]{vegawidget::vegawidget()}}}
}
\description{
Brushed Scatter Plot
}
\examples{
limn_xy(mtcars, mpg, hp, factor(cyl))
}
