% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter.R
\name{limn_xycol}
\alias{limn_xycol}
\title{Dynamic colored scatter plot}
\usage{
limn_xycol(.data, x, y, colors, ...)
}
\arguments{
\item{.data}{input data.frame}

\item{x, y}{bare symbols or expressions using columns from \code{.data} to map
to the x and y aesthetics of a scatter plot}

\item{colors}{a single unqouted expression of columns in \code{.data} that will be
dynamically bound to the colour aesthetic.}

\item{...}{other options passed to \code{\link[vegawidget:vegawidget]{vegawidget::vegawidget()}}}
}
\description{
Dynamic colored scatter plot
}
\details{
Note that this assumes that colors are of the same type...
}
