% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{sneezy}
\alias{sneezy}
\title{Exploring t-SNE embeddings interactively}
\usage{
sneezy(data, embedding, colour = NULL, max_bases)
}
\arguments{
\item{data}{a data.frame or matrix}

\item{embedding}{list output from \code{Rtsne::Rtsne()}}

\item{colour}{a column name in \code{data} corresponding to categorical variable}

\item{max_bases}{number of bases for grand tour (see \code{tourr::save_history()})}

\item{...}{other stuff I want to include.}
}
\value{
a shiny app
}
\description{
Exploring t-SNE embeddings interactively
}
\details{
Produces a 2 by 2 panel layout, where the upper left quadrant contains
2-d projections from a grand tour (1), the upper right quadrant contains
the 2-d t-SNE embedding (2), the lower left quadrant contains strip plots
of each rescaled variable in \code{data} (3), the lower right quadrant contains
a scatter plot of distance in embedding space against distance in the data
space (4).

UI elements and interactions

A range slider is bound to the side bar panel, with a play button which
will animate the projections in (1).

There are two brushing selections available in (2): a mouse drag
on (2) will highlight selected points on (1) and (2) and (3). This
is useful for both seeing how clusters that form in embedding space move
in high-dimensional space and what variables define the cluster from
(3). Pressing shift with a mouse drag will also highlight selected points
but this time revealing the simplex constructed on the embedding space
and how this simplex appears in high-dimensional space.

The scatter plot (4) of inter point distances, can be used for checking
spurious embeddings...

The strip plots can be used for highlighting point ranges ...
}
