% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tour-utils.R
\name{clamp}
\alias{clamp}
\alias{clamp_robust}
\alias{clamp_sd}
\title{Rescale all columns of a matrix}
\usage{
clamp(.data)

clamp_robust(.data)

clamp_sd(.data, sd = 1e-04)
}
\arguments{
\item{.data}{A numeric matrix}

\item{sd}{the value of each columns standard devation}
}
\description{
Rescale all columns of a matrix
}
\details{
These functions are used internally by the tour to rescale all
columns of \code{.data}.
\itemize{
\item \code{clamp()} rescales so all values for each column lie in the unit interval
\item \code{clamp_robust()} rescales by first centering by the median and then scaling
by MADs
\item \code{clamp_sd()} rescales all columns to have a fixed standard deviation.
}
}
\examples{
mv <- matrix(rnorm(300), ncol = 3)

clamp(mv)

clamp_robust(mv)

clamp_sd(mv)

}
