% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tour-utils.R
\name{compute_proj_dist}
\alias{compute_proj_dist}
\title{Compute Frobenius norm of matrix-like objects x and y}
\usage{
compute_proj_dist(x, y)
}
\arguments{
\item{x, y}{'matrix' like objects}
}
\description{
Compute Frobenius norm of matrix-like objects x and y
}
