% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liminal-package.R
\docType{package}
\name{liminal-package}
\alias{liminal}
\alias{liminal-package}
\title{liminal: interactive high-dimensional data visualisation}
\description{
For more details on the features and usage of liminal, see the vignettes:
\code{browseVignettes(package = "liminal")}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/sa-lee/liminal}
  \item \url{https://sa-lee.github.io/liminal}
  \item Report bugs at \url{https://github.com/sa-lee/liminal/issues/}
}

}
\author{
\strong{Maintainer}: Stuart Lee \email{stuart.andrew.lee@gmail.com} (\href{https://orcid.org/0000-0003-1179-8436}{ORCID}) [copyright holder]

}
\keyword{internal}
