% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tour-base.R
\name{limn_tour}
\alias{limn_tour}
\title{Tour a high dimensional dataset}
\usage{
limn_tour(
  .data,
  cols,
  color = NULL,
  tour_path = tourr::grand_tour(),
  rescale = clamp,
  morph = morph_center
)
}
\arguments{
\item{.data}{a data.frame to tour}

\item{cols}{Columns to tour. This can use a tidyselect specification.}

\item{color}{A variable name in \code{.data}, mapping to the color aesthetic. Default
is to not colour any points.}

\item{tour_path}{the tour path to take, the default is \code{\link[tourr:grand_tour]{tourr::grand_tour()}}.}

\item{rescale}{A function that rescales tour columns. Default is \code{\link[=clamp]{clamp()}}
To not perform any scaling use \code{\link[=identity]{identity()}}.}

\item{morph}{A callback function that modifies the projection, default is to
center the projection using \code{\link[=morph_center]{morph_center()}}.}
}
\description{
Tour a high dimensional dataset
}
