% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tour-base.R
\name{limn_tour}
\alias{limn_tour}
\title{Tour a high dimensional dataset}
\usage{
limn_tour(
  .data,
  cols,
  color = NULL,
  tour_path = tourr::grand_tour(),
  rescale = clamp,
  morph = "center"
)
}
\arguments{
\item{.data}{a data.frame to tour}

\item{cols}{Columns to tour. This can use a tidyselect specification
such as \code{\link[tidyselect:starts_with]{tidyselect::starts_with()}}.}

\item{color}{A variable name in \code{.data}, mapping to the color aesthetic, if
NULL points will be colored black.}

\item{tour_path}{the tour path to take, the default is \code{\link[tourr:grand_tour]{tourr::grand_tour()}}
but also works with \code{\link[tourr:guided_tour]{tourr::guided_tour()}}.}

\item{rescale}{A function that rescales \code{cols}, the default is to
\code{\link[=clamp]{clamp()}} the data to lie in the hyperdimensional unit cube. To not perform
any scaling use \code{\link[=identity]{identity()}}.}

\item{morph}{One of \code{c("center", "centre", "identity", "radial")}
that rescales each projection along the tour path. The default
is to center the projections and divide by half range. See \code{\link[=morph_center]{morph_center()}}
for details.}
}
\value{
A shiny app object
}
\description{
Tour a high dimensional dataset
}
\details{
The tour interface consists of two views
\enumerate{
\item the tour view which is a dynamic scatterplot
\item the axis view which shows the direction and magnitude of the
basis vectors being generated.
There are several other user controls available:
}
\itemize{
\item There is a play button, that when pressed will start the tour.
\item There is also a text view of the half range which is the maximum squared
Euclidean distance between points in the tour view. The half range
is a scale factor for projections and can be thought of as a way
of zooming in and out on points. It can be dynamically modified by scrolling
(via a mouse-wheel). To reset double click the tour view.
\item The legend can be toggled to highlight groups of points with
shift+mouse-click. Multiple groups can be selected in this way. To
reset double click the legend title.
}
}
\examples{
\dontrun{
# tour the first ten columns of the fake tree data
# loads the default interface
limn_tour(fake_trees, dim1:dim10)
# perform the same action but now coloring points
limn_tour(fake_trees, dim1:dim10, color = branches)
}

}
\seealso{
\code{\link[=compute_half_range]{compute_half_range()}},\code{\link[=limn_tour_xylink]{limn_tour_xylink()}}
}
