% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tour-linked-scatter.R
\name{limn_tour_xylink}
\alias{limn_tour_xylink}
\title{View a tour with external context}
\usage{
limn_tour_xylink(
  x,
  y,
  by = "rowid",
  x_color = NULL,
  y_color = NULL,
  tour_path = tourr::grand_tour(),
  rescale = clamp,
  morph = morph_center
)
}
\arguments{
\item{x}{a \code{data.frame} or \code{tibble} to tour}

\item{y}{a \code{data.frame} to link to \code{x}, representing a scatter plot}

\item{by}{a column (or columns) for defining how to join. See \code{dplyr::inner_join()}
for details. By default we use the row index for both tables.}

\item{x_color}{an optional bare column name in \code{x}, for the color mapping in the tour view}

\item{y_color}{an optional bare column name for the colour mapping the linked view}

\item{tour_path}{the tour path to take, the default is \code{\link[tourr:grand_tour]{tourr::grand_tour()}}.}

\item{rescale}{A function that rescales tour columns. Default is \code{\link[=clamp]{clamp()}}
To not perform any scaling use \code{\link[=identity]{identity()}}.}

\item{morph}{A callback function that modifies the projection, default is to
center the projection using \code{\link[=morph_center]{morph_center()}}.}
}
\description{
View a tour with external context
}
