% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tour-linked-scatter.R
\name{limn_tour_xylink}
\alias{limn_tour_xylink}
\title{Tour view linked}
\usage{
limn_tour_xylink(
  x,
  y,
  by = "rowid",
  x_color = NULL,
  y_color = NULL,
  tour_path = tourr::grand_tour(),
  clamp = TRUE,
  transformer = function(x) scale(x, scale = FALSE),
  ...
)
}
\arguments{
\item{x}{a \code{data.frame} or \code{tibble} to tour}

\item{y}{a \code{data.frame} to link to \code{x}, representing a scatter plot}

\item{by}{a column (or columns) for defining how to join. See \code{dplyr::inner_join()}
for details. By default we use the row index for both tables.}

\item{x_color}{an optional bare column name in \code{x}, for the color mapping in the tour view}

\item{y_color}{an optional bare column name for the colour mapping the linked view}

\item{tour_path}{the name of the tour path, defaul is \code{tourr::grand_tour()}}

\item{clamp}{rescale tour variables to lie in closed interval between 0 and 1.}

\item{...}{}
}
\description{
Tour view linked
}
