% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morphs.R
\name{morph_center}
\alias{morph_center}
\alias{morph_identity}
\alias{morph_radial}
\title{Morphing Projections}
\usage{
morph_center(proj, half_range)

morph_identity(proj, half_range)

morph_radial(proj, half_range, p_eff)
}
\arguments{
\item{proj}{a projection}

\item{half_range}{scale factor for projection}

\item{p_eff}{Effective dimensionality of reference data set.}
}
\description{
Morphing Projections
}
\examples{
proj <- matrix(rnorm(20), ncol = 2)
half_range <- compute_half_range(proj)
morph_center(proj, half_range)
morph_identity(proj, half_range)
morph_radial(proj, half_range, p_eff = 2)
}
