% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liminal-data.R
\name{limn_pal_tableau10}
\alias{limn_pal_tableau10}
\alias{limn_pal_tableau20}
\title{liminal color palettes}
\usage{
limn_pal_tableau10()

limn_pal_tableau20()
}
\value{
A character vector of hex color codes
}
\description{
liminal color palettes
}
\details{
Vectors of colors based of the schemes available in Vega-Lite.
Their main purpose is so you can use these palettes in \code{ggplot2} graphics.
}
\seealso{
https://vega.github.io/vega/docs/schemes/
}
